rename file, isCompatibleVersion takes only one argument, export versio string, rm test
