+++ packages/dht/src/connection/connectivityChecker.ts
@@ -8,7 +8,7 @@ import {
 import { ConnectionEvents, IConnection } from './IConnection'
 import { ClientWebsocket } from './websocket/ClientWebsocket'
 import { connectivityMethodToWebsocketUrl } from './websocket/WebsocketConnector'
-import { isCompatibleVersion } from '../helpers/versionCompatibility'
+import { isCompatibleVersion } from '../helpers/version'
 
 const logger = new Logger(module)
 
@@ -37,8 +37,7 @@ const CONNECTIVITY_CHECKER_TIMEOUT = 5000
 
 export const sendConnectivityRequest = async (
     request: ConnectivityRequest,
-    entryPoint: PeerDescriptor,
-    localVersion: string
+    entryPoint: PeerDescriptor
 ): Promise<ConnectivityResponse> => {
     let outgoingConnection: IConnection
     const wsServerInfo = {
@@ -83,7 +82,7 @@ export const sendConnectivityRequest = async (
                         const remoteVersion = connectivityResponseMessage.version
                         outgoingConnection!.off('data', listener)
                         clearTimeout(timeoutId)
-                        if (isCompatibleVersion(localVersion, remoteVersion)) {
+                        if (isCompatibleVersion(remoteVersion)) {
                             resolve(connectivityResponseMessage)
                         } else {
                             reject(`Invalid version: ${remoteVersion}`)
