+++ packages/dht/src/connection/webrtc/WebrtcConnectorRpcLocal.ts
@@ -18,8 +18,7 @@ import { ManagedWebrtcConnection } from '../ManagedWebrtcConnection'
 import { NodeWebrtcConnection } from './NodeWebrtcConnection'
 import { WebrtcConnectorRpcRemote } from './WebrtcConnectorRpcRemote'
 import { DhtAddress, getNodeIdFromPeerDescriptor } from '../../identifiers'
-import { version as localVersion } from '../../../package.json'
-import { isCompatibleVersion } from '../../helpers/versionCompatibility'
+import { isCompatibleVersion } from '../../helpers/version'
 import { ConnectionID } from '../IConnection'
 
 const logger = new Logger(module)
@@ -88,7 +87,7 @@ export class WebrtcConnectorRpcLocal implements IWebrtcConnectorRpc {
             if (this.config.ongoingConnectAttempts.has(nodeId)) {
                 this.config.ongoingConnectAttempts.delete(nodeId)
             }
-            if (!isCompatibleVersion(localVersion, remoteVersion)) {
+            if (!isCompatibleVersion(remoteVersion)) {
                 managedConnection!.rejectHandshake(HandshakeError.UNSUPPORTED_VERSION)
             } else {
                 managedConnection!.acceptHandshake()
