+++ packages/dht/src/connection/websocket/WebsocketConnector.ts
@@ -29,8 +29,7 @@ import { DISABLE_CONNECTIVITY_PROBE, attachConnectivityRequestHandler } from '..
 import * as Err from '../../helpers/errors'
 import { Empty } from '../../proto/google/protobuf/empty'
 import { DhtAddress, areEqualPeerDescriptors, getNodeIdFromPeerDescriptor } from '../../identifiers'
-import { version as localVersion } from '../../../package.json'
-import { isCompatibleVersion } from '../../helpers/versionCompatibility'
+import { LOCAL_PROTOCOL_VERSION, isCompatibleVersion } from '../../helpers/version'
 
 const logger = new Logger(module)
 
@@ -175,7 +174,7 @@ export class WebsocketConnector {
                 host: '127.0.0.1',
                 natType: NatType.UNKNOWN,
                 ipAddress: ipv4ToNumber('127.0.0.1'),
-                version: localVersion
+                version: LOCAL_PROTOCOL_VERSION
             }
         }
         if (!this.config.entrypoints || this.config.entrypoints.length === 0) {
@@ -190,7 +189,7 @@ export class WebsocketConnector {
                 },
                 // TODO: Resolve the given host name or or use as is if IP was given. 
                 ipAddress: ipv4ToNumber('127.0.0.1'),
-                version: localVersion
+                version: LOCAL_PROTOCOL_VERSION
             }
         }
         for (const reattempt of range(ENTRY_POINT_CONNECTION_ATTEMPTS)) {
@@ -204,7 +203,7 @@ export class WebsocketConnector {
                     selfSigned
                 }
                 if (!this.abortController.signal.aborted) {
-                    return await sendConnectivityRequest(connectivityRequest, entryPoint, localVersion)
+                    return await sendConnectivityRequest(connectivityRequest, entryPoint)
                 } else {
                     throw new Err.ConnectionFailed('ConnectivityChecker is destroyed')
                 }
@@ -306,7 +305,7 @@ export class WebsocketConnector {
         const nodeId = getNodeIdFromPeerDescriptor(sourcePeerDescriptor)
         if (this.ongoingConnectRequests.has(nodeId)) {
             const ongoingConnectRequest = this.ongoingConnectRequests.get(nodeId)!
-            if (!isCompatibleVersion(localVersion, remoteVersion)) {
+            if (!isCompatibleVersion(remoteVersion)) {
                 ongoingConnectRequest.rejectHandshake(HandshakeError.UNSUPPORTED_VERSION)
             } else {
                 ongoingConnectRequest.attachImplementation(websocketServerConnection)
@@ -322,7 +321,7 @@ export class WebsocketConnector {
                 targetPeerDescriptor
             )
             managedConnection.setRemotePeerDescriptor(sourcePeerDescriptor)
-            if (!isCompatibleVersion(localVersion, remoteVersion)) {
+            if (!isCompatibleVersion(remoteVersion)) {
                 managedConnection.rejectHandshake(HandshakeError.UNSUPPORTED_VERSION)
             } else if (targetPeerDescriptor && !areEqualPeerDescriptors(this.localPeerDescriptor!, targetPeerDescriptor)) {
                 managedConnection.rejectHandshake(HandshakeError.INVALID_TARGET_PEER_DESCRIPTOR)
