+++ packages/dht/src/exports.ts
@@ -26,3 +26,4 @@ export {
     areEqualPeerDescriptors,
     getNodeIdFromPeerDescriptor
 } from './identifiers'
+export { LOCAL_PROTOCOL_VERSION } from './helpers/version'
